+++ starter-kits/nextjs-starter-kit/layouts/Document/index.ts
@@ -1,1 +1,2 @@
 export * from "./Document";
+export * from "./DocumentProviders";